requirements.txt
flask
mongoengine
redis
qcloud-cos-python
requests
PyYAML
flask-cors
